fixed linter errors
