+++ apps/frontend/src/contexts/AuthContextValue.ts
@@ -1,0 +1,21 @@
+import { createContext } from 'react';
+import {
+  User,
+  LoginCredentials,
+  RegisterCredentials,
+  AuthError,
+} from '../types/auth';
+
+export interface AuthContextType {
+  user: User | null;
+  isAuthenticated: boolean;
+  isLoading: boolean;
+  error: AuthError | null;
+  login: (credentials: LoginCredentials) => Promise<void>;
+  register: (credentials: RegisterCredentials) => Promise<void>;
+  logout: () => Promise<void>;
+}
+
+export const AuthContext = createContext<AuthContextType | undefined>(
+  undefined
+);
